+++ autocompleteDataset.js
@@ -1,3 +1,8 @@
+// we need to export using commonjs for ease of usage in all
+// JavaScript environments
+
+/* eslint-disable import/no-commonjs */
+
 import createAutocompleteDataset from './src/createAutocompleteDataset.js';
 import css from './src/places.scss';
 import insertCss from 'insert-css';