+++ scripts/build.sh
@@ -20,14 +20,17 @@ for bundle in "${bundles[@]}"
 do
   dist_file="$dist_dir_cdn/${bundle}.js"
   dist_file_min="$dist_dir_cdn/${bundle}.min.js"
-  dist_file_sourcemap="$dist_dir_cdn/${bundle}.js.map"
-  dist_file_sourcemap_min="$dist_dir_cdn/${bundle}.min.js.map"
+  source_map="${bundle}.js.map"
+  source_map_min="${bundle}.min.js.map"
+  dist_file_sourcemap="$dist_dir_cdn/${source_map}"
+  dist_file_sourcemap_min="$dist_dir_cdn/${source_map_min}"
 
   echo "$license" | cat - "${dist_file}" > /tmp/out && mv /tmp/out "${dist_file}"
 
   uglifyjs "${dist_file}" \
     --in-source-map "${dist_file_sourcemap}" \
     --source-map "${dist_file_sourcemap_min}" \
+    --source-map-url "${source_map_min}" \
     --preamble "$license" \
     -c warnings=false \
     -m \
@@ -44,4 +47,4 @@ done
 BABEL_DISABLE_CACHE=1 BABEL_ENV=npm babel -q index.js -o "$dist_dir/index.js"
 BABEL_DISABLE_CACHE=1 BABEL_ENV=npm babel -q autocompleteDataset.js -o "$dist_dir/autocompleteDataset.js"
 BABEL_DISABLE_CACHE=1 BABEL_ENV=npm babel -q instantsearchWidget.js -o "$dist_dir/instantsearchWidget.js"
-BABEL_DISABLE_CACHE=1 BABEL_ENV=npm babel -q --out-dir "$dist_dir/src/" src/
+BABEL_DISABLE_CACHE=1 BABEL_ENV=npm babel -q src/ --out-dir "$dist_dir/src/" --ignore test.js