+++ src/__mocks__/algoliasearch/lite.js
@@ -1,0 +1,22 @@
+const defaultSearchStub = jest.fn(() => Promise.resolve());
+let searchStub = defaultSearchStub;
+const addAlgoliaAgent = jest.fn();
+const search = jest.fn((...args) => searchStub(...args));
+
+const algoliasearch = {
+  initPlaces: jest.fn(() =>
+    ({
+      as: {
+        addAlgoliaAgent
+      },
+      search
+    })
+  )
+};
+
+algoliasearch.__searchSpy = search;
+algoliasearch.__addAlgoliaAgentSpy = addAlgoliaAgent;
+algoliasearch.__setSearchStub = fn => { searchStub = fn; };
+algoliasearch.__clearSearchStub = () => { searchStub = defaultSearchStub; };
+
+export default algoliasearch;