+++ src/createAutocompleteSource.js
@@ -6,6 +6,7 @@ export default function createAutocompleteSource({
   clientOptions,
   apiKey,
   appId,
+  hitsPerPage,
   aroundLatLng,
   aroundRadius,
   aroundLatLngViaIP,
@@ -15,6 +16,8 @@ export default function createAutocompleteSource({
   useDeviceLocation = false,
   language = navigator.language.split('-')[0],
   onHits = () => {},
+  onError = e => { throw e; },
+  onRateLimitReached,
   type
 }) {
   const placesClient = algoliasearch.initPlaces(
@@ -24,16 +27,16 @@ export default function createAutocompleteSource({
   );
   placesClient.as.addAlgoliaAgent(`Algolia Places ${version}`);
 
-  let defaultQueryParams = {
+  const defaultQueryParams = {
     countries,
-    hitsPerPage: 5,
+    hitsPerPage: hitsPerPage || 5,
     language,
     type
   };
 
   if (aroundLatLng) {
     defaultQueryParams.aroundLatLng = aroundLatLng;
-  } else if (typeof aroundLatLngViaIP !== 'undefined') {
+  } else if (aroundLatLngViaIP !== undefined) {
     defaultQueryParams.aroundLatLngViaIP = aroundLatLngViaIP;
   }
 
@@ -44,7 +47,7 @@ export default function createAutocompleteSource({
   let userCoords;
   if (useDeviceLocation) {
     navigator.geolocation.watchPosition(
-      ({coords}) => userCoords = `${coords.latitude},${coords.longitude}`
+      ({coords}) => { userCoords = `${coords.latitude},${coords.longitude}`; }
     );
   }
 
@@ -56,15 +59,15 @@ export default function createAutocompleteSource({
       }))
       .then(
         content => {
-          const hits = content.hits.map((hit, hitIndex) => {
-            return formatHit({
+          const hits = content.hits.map((hit, hitIndex) =>
+            formatHit({
               formatInputValue,
               hit,
               hitIndex,
               query,
               rawAnswer: content
-            });
-          });
+            })
+          );
 
           onHits({
             hits,
@@ -75,5 +78,13 @@ export default function createAutocompleteSource({
           return hits;
         }
       )
-      .then(cb);
+      .then(cb)
+      .catch(e => {
+        if (e.statusCode === 429) {
+          onRateLimitReached();
+          return;
+        }
+
+        onError(e);
+      });
 }