+++ src/findCountryCode.test.js
@@ -1,0 +1,20 @@
+jest.unmock('./findCountryCode.js');
+import findCountryCode from './findCountryCode.js';
+
+describe('findCountryCode', () => {
+  const testCases = [{
+    name: 'empty array',
+    input: [],
+    expected: undefined
+  }, {
+    name: 'match',
+    input: ['country/us', 'country/fr'],
+    expected: 'us'
+  }];
+
+  testCases.forEach(
+    testCase => it(`${testCase.name} test case`, () =>
+      expect(findCountryCode(testCase.input)).toEqual(testCase.expected)
+    )
+  );
+});